requirements.txt
pandas 
numpy 
PyMuPDF
sentence-transformers
faiss-cpu
google-generativeai
dotenv
streamlit
python-docx